# config/swagger/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/register/guest":{"post":{"operationId":"V1RegisterGuestHttpController_execute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/V1RegisterGuestHttpRequest"}}}},"responses":{"201":{"description":""}}}},"/api/v1/register/member":{"post":{"operationId":"V1RegisterMemberHttpController_execute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/V1RegisterMemberHttpRequest"}}}},"responses":{"201":{"description":""}}}},"/api/v1/register/admin":{"post":{"operationId":"V1RegisterAdminHttpController_execute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/V1RegisterAdminHttpRequest"}}}},"responses":{"201":{"description":""}},"security":[{"x-api-key":[]}]}},"/api/v1/sign-in":{"post":{"operationId":"V1SignInHttpController_execute","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/V1SignInHttpRequest"}}}},"responses":{"200":{"description":""}}}},"/api/v1/profile":{"get":{"operationId":"V1GetProfileHttpController_execute","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}}},"info":{"title":"User Managemenet API","description":"The User Management API description","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"x-api-key":{"type":"apiKey","in":"header","name":"x-api-key"}},"schemas":{"V1RegisterGuestHttpRequest":{"type":"object","properties":{}},"V1RegisterMemberHttpRequest":{"type":"object","properties":{"username":{"type":"string","example":"username","description":"Username"},"password":{"type":"string","example":"password","description":"Password"},"email":{"type":"string","example":"email@example.com","description":"Email"},"firstName":{"type":"string","example":"firstName","description":"First Name"},"lastName":{"type":"string","example":"lastName","description":"Last Name"}},"required":["username","password","email","firstName","lastName"]},"V1RegisterAdminHttpRequest":{"type":"object","properties":{"username":{"type":"string","example":"username"},"email":{"type":"string","example":"email@example.com"},"password":{"type":"string","example":"password"},"firstName":{"type":"string","example":"firstName"},"lastName":{"type":"string","example":"lastName"}},"required":["username","email","password","firstName","lastName"]},"V1SignInHttpRequest":{"type":"object","properties":{"username":{"type":"string","example":"username","description":"Username"},"email":{"type":"string","example":"email@example.com","description":"Email"},"password":{"type":"string","example":"password","description":"Password"}},"required":["username","email","password"]}}}}